# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{userId}/permissions": {
      "get": {
        "operationId": "UserPermissionsController_get",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSetType"
                }
              }
            }
          }
        },
        "tags": [
          "UserPermissions"
        ]
      },
      "put": {
        "operationId": "UserPermissionsController_set",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionSetInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSetType"
                }
              }
            }
          }
        },
        "tags": [
          "UserPermissions"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/session": {
      "get": {
        "operationId": "AuthController_ownSessions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "offset": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/session/{userId}": {
      "get": {
        "operationId": "AuthController_userSessions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "offset": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/session/{userId}/{sid}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/report/history": {
      "get": {
        "operationId": "ReportController_history",
        "parameters": [
          {
            "name": "term",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceHistoryType"
                }
              }
            }
          }
        },
        "tags": [
          "Report"
        ]
      }
    },
    "/report/{kind}": {
      "get": {
        "operationId": "ReportController_report",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "enum": [
                "best",
                "catalog",
                "drops",
                "low",
                "new"
              ]
            }
          },
          {
            "name": "stores",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minVolume",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxVolume",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "flavors",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "excludeFlavors",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "types",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "countries",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "minDiscount",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "window",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "yesterday",
                "week",
                "month",
                "year"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "abv",
                "age",
                "countryName",
                "discountPct",
                "name",
                "previousPrice",
                "price",
                "storeName",
                "type",
                "volumeMl"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "enum": [
                50,
                100,
                150,
                200
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReportRowType"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "offset": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Report"
        ]
      }
    },
...
    "/store/{slug}": {
      "get": {
        "operationId": "StoreController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDetailType"
                }
              }
            }
          }
        },
        "tags": [
          "Store"
        ]
      },
      "patch": {
        "operationId": "StoreController_setActive",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListItemType"
                }
              }
            }
          }
        },
        "tags": [
          "Store"
        ]
      }
    },
...
  },
  "info": {
    "title": "Whisky Scrapper",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PermissionGroupType": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "enum": [
              "authenticated",
              "permission",
              "product",
              "public",
              "self",
              "session",
              "store",
              "user"
            ]
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "create",
                "read",
                "update",
                "edit",
                "delete",
                "list",
                "logout",
                "read_permissions",
                "set_permissions"
              ]
            }
          }
        },
        "required": [
          "resource",
          "actions"
        ]
      },
      "PermissionSetType": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionGroupType"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "PermissionGroupInputDto": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "enum": [
              "authenticated",
              "permission",
              "product",
              "public",
              "self",
              "session",
              "store",
              "user"
            ]
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "create",
                "read",
                "update",
                "edit",
                "delete",
                "list",
                "logout",
                "read_permissions",
                "set_permissions"
              ]
            }
          }
        },
        "required": [
          "resource",
          "actions"
        ]
      },
      "PermissionSetInputDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionGroupInputDto"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "AuthLoginDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "AccessToken": {
        "type": "object",
        "properties": {
          "access": {
            "type": "string"
          }
        },
        "required": [
          "access"
        ]
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "sid": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "expires": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "sid",
          "ip",
          "userAgent",
          "expires"
        ]
      },
      "ReportRowType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "nameOrig": {
            "type": "string"
          },
          "age": {
            "type": "number",
            "nullable": true
          },
          "abv": {
            "type": "number",
            "nullable": true
          },
          "volumeMl": {
            "type": "number",
            "nullable": true
          },
          "storeSlug": {
            "type": "string"
          },
          "storeName": {
            "type": "string"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "countryName": {
            "type": "string",
            "nullable": true
          },
          "countryIcon": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number"
          },
          "oldPrice": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "inStock": {
            "type": "boolean"
          },
          "promo": {
            "type": "boolean"
          },
          "previousPrice": {
            "type": "number",
            "nullable": true
          },
          "referencePrice": {
            "type": "number",
            "nullable": true
          },
          "discountPct": {
            "type": "number",
            "nullable": true
          },
          "isNew": {
            "type": "boolean"
          },
          "daysNew": {
            "type": "number",
            "nullable": true
          },
          "flavors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "firstSeen": {
            "type": "string"
          },
          "capturedDate": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "url",
          "name",
          "nameOrig",
          "age",
          "abv",
          "volumeMl",
          "storeSlug",
          "storeName",
          "brand",
          "type",
          "countryCode",
          "countryName",
          "countryIcon",
          "price",
          "oldPrice",
          "currency",
          "inStock",
          "promo",
          "previousPrice",
          "referencePrice",
          "discountPct",
          "isNew",
          "daysNew",
          "flavors",
          "firstSeen",
          "capturedDate"
        ]
      },
      "PriceHistoryPointType": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "date",
          "price"
        ]
      },
      "PriceHistoryType": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/ReportRowType"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceHistoryPointType"
            }
          }
        },
        "required": [
          "product",
          "series"
        ]
      },
...
      "StoreListItemType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "tier": {
            "type": "number",
            "nullable": true
          },
          "needsBrowser": {
            "type": "boolean",
            "nullable": true
          },
          "retailChain": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "lastSuccessfulSyncAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "baseUrl",
          "color",
          "active",
          "tier",
          "needsBrowser",
          "retailChain",
          "category",
          "lastSuccessfulSyncAt"
        ]
      },
      "SyncLogType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "added": {
            "type": "number"
          },
          "removed": {
            "type": "number"
          },
          "updated": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "storeId",
          "added",
          "removed",
          "updated",
          "total"
        ]
      },
      "StoreDetailType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "tier": {
            "type": "number",
            "nullable": true
          },
          "needsBrowser": {
            "type": "boolean",
            "nullable": true
          },
          "retailChain": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "lastSuccessfulSyncAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "productCount": {
            "type": "number"
          },
          "lastSync": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SyncLogType"
              }
            ]
          },
          "recentSyncs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncLogType"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "baseUrl",
          "color",
          "active",
          "tier",
          "needsBrowser",
          "retailChain",
          "category",
          "lastSuccessfulSyncAt",
          "createdAt",
          "productCount",
          "lastSync",
          "recentSyncs"
        ]
      },
      "StoreActiveDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ]
      },
...
    }
  }
}